requirements.txt
pip==24.3.1
gunicorn==22.0.0
Flask==3.1.0
Jinja2==3.1.4
numpy==1.26.4
pillow==10.4.0
